services/auth_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==4.0.0
pika==1.3.1
pyjwt==2.6.0
gunicorn==20.1.0